Rework how the cards are created and dealt Cards are now taken from a vec with values from 1 to 52. Each value represents a card. Suits are represented by the value of the card % 13. Every suit owns 13 cards.
Cards are now created and formatted in the Card's new() function. Only the value is passed to create the card and the function assigns the suit.

Card struct now implements Display, Debug and PartialEq. They are displayed using the standard
"{rank} of {suit}"· Debug and PartialEq are used for testing.