Change layout to display hand results on screen ui.rs ----- After pressing enter to change cards, the program resets the deck and shows what combination the hand contained, if any. The points given are also shown.
Cards and ascii cards are always show, even when a game isn't active.
Only the points gained aren't shown during a game.

Added render_message() to display the combination found in the hand and
the points it gave.

Added tests for each display option currently available.

lib.rs
------
Created the reset_deck() function which inserts values from the
hand and the discarded pile back into the deck. This is to ensure that
the deck has the same values every time cards are dealt.

Added tests for the change_cards(), reset_deck(), and deal() functions.

Fixed card checking tests. Cards were missing the rank field when
created.

Deleted the Display implementation for cards. It was no longer used.
Deleted the unused fmt import.

Changed the name of the hand checking functions for easier testing of
only those. They now all start with "hand_".